feat: add MCP tool annotations, extract shared helpers, fix tag/guard issues
- Add readOnlyHint, destructiveHint, idempotentHint annotations to all 23 tools
- Extract shared helpers (_get_jira, _get_confluence, _check_write, _ok, _err)
  to servers/_helpers.py — eliminates duplication across 4 server modules
- Fix jira_download_attachment: tag changed from "read" to "write" (writes to
  local disk), add _check_write guard for read-only mode enforcement
- Fix proprietary data in docstring/tests: use generic customfield_12345/MyTeam
- Update AGENTS.md with MCP compliance rules, correct tool count to 23
- Add .claude/, .cursor/, .cursorrules to .gitignore
