+++ .gitignore
@@ -32,3 +32,6 @@ Thumbs.db
 .idea/
 .vscode/
 *.iml
+.claude/
+.cursorrules
+.cursor/
